fix calculation of ROS topic subscriber count
if a subscriber disconnected the node never noticed this since it always
does an `std::max` on the current value and the subscriber count of each
stream. this is meant to find the largest amount of subscribers over all
stream but incorrectly never resets it to a lower value if the stream
with the highest subscriber count had the now disconnected subscriber.

this is easiest to reproduce by having a single camera with a single
stream and connecting to it with `image_view`: upon first launch the
`camera_aravis2` node will write the debug message `Acquisition
started`, after stopping `image_view` it logs `Acquisition stopped`.
but when starting again `image_view` it no longer logs `Acquisition
start` without this fix and no image is written to the ROS topic.